Move caps and device spec utils to `oci` pkg
